Remove some unnecessary pub's, move storage implementations to separate files, some renames
