+++ crates/bitwarden/src/state/mod.rs
@@ -1,2 +1,4 @@
 pub mod state;
 pub(crate) mod state_service;
+
+mod storage;
